requirements.txt
streamlit
openai
pypdf
python-docx
python-pptx
openpyxl
python-dotenv
